refactor: simvars
this patch introduces a simvars.json config file. simvar controls are now persisted in the .ace/ project folder.

additionally, I have refactored the component and general React structure of the simvar system, removing a not-so-needed context and cleaned up the control editor, as well as normalized saving of simvar values.